[MR] | 2020.08.14 | back to develop

+++ backend/src/main/java/com/web/blog/service/CommentService.java
@@ -22,6 +22,8 @@ public class CommentService {
 
 	private final CommentRepository commentRepository;
 	private final UsersRepository userRepository;
+
+	private final String ADMIN = "ROLE_ADMIN";
 	
 	public void createComments(Comments comment) {
 		System.out.println(comment);
@@ -29,8 +31,8 @@ public class CommentService {
 				.pid(comment.getPid())
 				.comment(comment.getComment())
 				.email(comment.getEmail())
-				.comment_time(LocalDateTime.now())
-				.update_time(LocalDateTime.now())
+				.commentTime(LocalDateTime.now())
+				.updateTime(LocalDateTime.now())
 				.build());
 		// 댓글 작성 시 댓글 작성자 경험치 상승
 		Optional<Users> user = userRepository.findByEmail(comment.getEmail());
@@ -59,8 +61,14 @@ public class CommentService {
 	}
 	
 	@Transactional
-	public void deleteComments(int cmid) {
-		commentRepository.deleteByCmid(cmid);
+	public boolean deleteComments(String user, int cmid, String role) {
+		Comments comment = commentRepository.findByCmid(cmid);
+		if(user.equals(comment.getEmail()) | role.equals(ADMIN)) {
+			commentRepository.deleteByCmid(cmid);
+			return true;
+		}else {
+			return false;
+		}
 	}
 
 }
